api/requirements.txt
anyio==3.5.0
blinker==1.7.0
click==8.1.7
faker==24.7.1
flask==3.0.1
flask_cors==4.0.0
flask-httpauth==4.8.0
flask-jwt-extended==4.6.0
flask-marshmallow==1.2.1
flask-migrate==4.0.7
Flask-SQLAlchemy==3.1.1
freezegun==1.4.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
pandas==2.2.2
PyJWT==2.8.0
pytest==8.0.0
pytest-cov==5.0.0
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
SQLAlchemy==2.0.27
tensorflow==2.16.1
Werkzeug==3.0.1
xgboost==2.0.3
Flask-Migrate==4.0.7